+++ pkg/genericapiserver/BUILD
@@ -22,9 +22,7 @@ go_library(
         "hooks.go",
         "resource_config.go",
         "resource_encoding_config.go",
-        "reststorage_interfaces.go",
         "serve.go",
-        "services.go",
         "storage_factory.go",
         "tunneler.go",
     ],
